image_processor/requirements.txt
pymysql
django
djangorestframework
confluent-kafka
python-dotenv
cryptography
pymilvus
numpy
requests
pillow
dlib
aiohttp
networkx 
scipy
chinese_whispers 